+++ babel.config.js
@@ -1,0 +1,8 @@
+// Only used for testing with jest!
+module.exports = {
+  // Jest
+  presets: [
+    ['@babel/preset-env', {targets: {node: 'current'}}],
+     '@babel/preset-typescript',
+  ],
+};